doc: relnotes: Document the OOT Module eXample (OMX) structure
The OMX module provides an example for extending an existing
non-Zephyr repository to be recognized by the Zephyr build
system as an out-of-tree module. In order to avoid conflicts
with existing directory structures, build systems, and filenames
the Zephyr build and CI system interfacing files and directories
are put within the (already required) zephyr subdirectory. These
parallel the organization and naming patterns of Zephyr to align
with user and tool expectations.

Trivial versions of samples, subsystems, and test files have
been provided to demonstrate the scalable structure and verify
the build and CI processes maintain support for extensible
out-of-tree modules.  This is designed expecting the example
functionality will grow to include boards, DTS, drivers, and
tool and process support to maintain at least the same
capabilities as available with in-tree builds.

This example has been verified to work on Ubuntu 18.04
and west 0.8.0a1 with the following command sequence:

    west init \
        -m <path-to-repository>/oot_module_example \
        --mr <repository-branch> \
        --mf zephyr/manifests/zp/zephyr-v2.3.0/west.yml \
        gs \
 && cd gs \
 && west update \
 && source zephyr/zephyr-env.sh \
 && zephyr/scripts/sanitycheck -p native_posix -T oot_module_example/zephyr \
 && west build -p always -t guiconfig -b native_posix \
        zephyr/samples/hello_world \
 && cd oot_module_example/zephyr \
 && make htmldocs

It was discovered during this development that some of the needed
Zephyr build system functionality previously available is broken
in zephyr-v2.4.0.

Signed-off-by: Gregory Shue <gregory.shue@legrand.us>